feat: mariadb container now works properly

+++ src/requirements/mariadb/tools/init-db.sh
@@ -1,13 +1,21 @@
-#!/bin/sh
+#! /bin/bash
+
+if [ -f "$DB_DATADIR/mysql/user.frm" ]; then
+    mariadbd-safe --port=3306 --bind-address=0.0.0.0 --datadir=$DB_DATADIR
+    exit $?
+fi
 
 envsubst < /init.sql.template > /init.sql
 
-rc-service mariadb start
+mariadb-install-db --datadir=$DB_DATADIR --skip-test-db > /dev/null
 
-sleep 5
+mariadbd-safe --datadir=$DB_DATADIR --init-file=/init.sql --silent&
 
-mariadb < /init.sql
+echo "Waiting for mariadb to init..."
+until mariadb-admin -u root -p$DB_ROOT_PASSWORD shutdown --silent; do
+    sleep 1
+done
 
-mariadb-admin -u root -p$DB_ROOT_PASSWORD shutdown
+echo -e "Starting db"
 
-mariadbd-safe --port=3306 --bind-address=0.0.0.0 --datadir='/var/lib/mysql'
+mariadbd-safe --port=3306 --bind-address=0.0.0.0 --datadir=$DB_DATADIR
